pyproject.toml
[tool.poetry]
authors = ["Lunarmagpie <Bambolambo0@gmail.com>"]
description = "A docstring parser for hikari-crescent."
license = "MPL-2.0"
name = "crescent-ext-docstrings"
packages = [
  {include = "crescent/ext/docstrings" },
]
version = "0.1.0"

[tool.poetry.dependencies]
docstring-parser = "^0.14.1"
hikari-crescent = "^0.2.0"
python = ">=3.8,<3.11"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
codespell = "^2.1.0"
flake8 = "^4.0.1"
isort = "^5.10.1"
mypy = "^0.961"
nox = "^2022.1.7"

[tool.mypy]
strict = true
namespace_packages = true
warn_return_any = true
warn_redundant_casts = true

[tool.black]
line-length = 88
target-version = ['py38']
skip-magic-trailing-comma=true

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]
